% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnst-data.R
\docType{data}
\name{cnst}
\alias{cnst}
\title{Table with material constants used in model}
\format{An object of class \code{"data.frame"}}
\source{
J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999
}
\usage{
cnst
}
\description{
This table is reproduced from Tab. 1 in reference paper.
}
\examples{
cnst
}
\references{
"Theory and experiment of InGaAsP and AlInGaAs long-wavelength strained quantum-well lasers" J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999
(\href{http://ieeexplore.ieee.org/document/760325/}{IEEE})
}
\keyword{datasets}
