% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Eg2PL_f1}
\alias{Eg2PL_f1}
\title{Conversion of Bandgap into photoluminescence (PL) wavelength in nm}
\usage{
Eg2PL_f1(Eg)
}
\arguments{
\item{Eg}{bandgap in eV units}
}
\value{
PL in nm as double
}
\description{
Eg into PL in nm
}
