% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{func4root}
\alias{func4root}
\title{Function for which we want to find roots}
\usage{
func4root(x)
}
\arguments{
\item{x}{Ga molar fraction}
}
\description{
Function used with function PQ_X_EgStrain tp find root,
which would give us X Ga molar fraction
}
