% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LatticeConst}
\alias{LatticeConst}
\title{lattice constant from eps_xy and for InP substrate}
\usage{
LatticeConst(eps_xy)
}
\arguments{
\item{eps_xy}{Strain in plane of material in absolute units}
}
\value{
Lattice constant in nm units
}
\description{
details tbd
}
