% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_Eg_XY}
\alias{modPQ_Eg_XY}
\title{Vectorized Eg from X and Y for InGaAsP}
\usage{
modPQ_Eg_XY(x, y)
}
\arguments{
\item{x}{Ga molar fraction}

\item{y}{As molar fraction}
}
\description{
Using mapply to \code{PQ_Eg_XY}
}
