% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_EgStrain}
\alias{modPQ_EgStrain}
\title{Data.frame for plotting Eg vs Strain for InGaAsP}
\usage{
modPQ_EgStrain(x, y)
}
\arguments{
\item{x}{Ga molar fraction}

\item{y}{As molar fraction}
}
\description{
Combining output of eps_xy and PQ_Eg_XY functions
}
