% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_eps_xy_f1}
\alias{modPQ_eps_xy_f1}
\title{Vectorized Eg from X and Y for InGaAsP}
\usage{
modPQ_eps_xy_f1(x, y)
}
\arguments{
\item{x}{Ga molar fraction}

\item{y}{As molar fraction}
}
\description{
Using mapply to \code{PQ_Eg_XY}
}
