% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_PL2Eg}
\alias{modPQ_PL2Eg}
\title{Transform vectors of PL in nm into Eg in eV}
\usage{
modPQ_PL2Eg(PL)
}
\arguments{
\item{PL}{Ga molar fraction}
}
\description{
Transform vectors of PL in nm into Eg in eV
}
