% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_X_EgStrain}
\alias{modPQ_X_EgStrain}
\title{Vectorized Eg from X and Y for InGaAsP}
\usage{
modPQ_X_EgStrain(Eg, Strain)
}
\arguments{
\item{Eg}{bandgap Eg in eV}

\item{Strain}{Strain of layer}
}
\description{
Using mapply to \code{PQ_X_EgStrain}
}
