% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_XY_EgStrain}
\alias{modPQ_XY_EgStrain}
\title{Data.frame for plotting X vs Y for InGaAsP}
\usage{
modPQ_XY_EgStrain(Eg, Strain)
}
\arguments{
\item{Eg}{bandgap Eg in eV}

\item{Strain}{Strain of layer}
}
\description{
Combining output of \code{PQ_X_EgStrain} and PQ_Y_XStrain functions0
}
