% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPQ.R
\name{modPQ_Y_EgStrain}
\alias{modPQ_Y_EgStrain}
\title{Vectorized PQ_Y_EgStrain}
\usage{
modPQ_Y_EgStrain(x, Strain)
}
\arguments{
\item{x}{Ga molar fraction}

\item{Strain}{Strain of layer}
}
\description{
Vectorized function \code{PQ_Y_EgStrain}
}
