% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PL2Eg_f1}
\alias{PL2Eg_f1}
\title{Convertsion of PL wavelength to bandgap in eV}
\usage{
PL2Eg_f1(PL)
}
\arguments{
\item{PL}{photoluminescence wavelength in nm. See examples.}
}
\value{
Bandgap Eg as double in eV units
}
\description{
details tbd
}
\examples{
PL2Eg_f1(1200) # 1.033202 eV
}
