% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PQ_Eg_unstrained}
\alias{PQ_Eg_unstrained}
\title{Unstrained bandgap Eg in eV from polynomial eq.3}
\usage{
PQ_Eg_unstrained(x, y)
}
\arguments{
\item{x}{Ga molar fraction}

\item{y}{As molar fraction}
}
\description{
Strain effects are not included, Eg in eV from polynomial eq.3
}
