% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PQ_eps_xy_f1}
\alias{PQ_eps_xy_f1}
\title{InGaAsP Strain in absolute units from X and Y}
\usage{
PQ_eps_xy_f1(x, y)
}
\arguments{
\item{x}{Ga molar fraction}

\item{y}{As molar fraction}
}
\description{
Function use interpolation PQ_interp_a0
}
