% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PQ_X_YStrain}
\alias{PQ_X_YStrain}
\title{Finding X molar fraction from Y molar fraction and Strain}
\usage{
PQ_X_YStrain(y, strain)
}
\arguments{
\item{y}{As molar fraction}

\item{strain}{Strain in absolute units}
}
\value{
double, x (Ga) molar fraction
}
\description{
Solving interpolation function from eq.1
a0 lattice constant of compound calculated from strain, and given y molar fraction
allows to find X molar fraction of compound.
Usually used to find eg. lattice matched InGaAs composition Y=1, and strain=0
}
