% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1a-data.R
\docType{data}
\name{fig1a}
\alias{fig1a}
\title{InGaAsP X vs Y for fixed strain or bandgap}
\format{\code{"data.frame"}}
\source{
J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999

#' @examples
data("fig1a")
head(fig1a)
}
\usage{
fig1a
}
\description{
Relation between InGaAsP X and Y molar fractions digitized from Fig. 1(a) in reference paper.
Data was digitized using Engauge Digitizer from Fig. 1(a) to compare against R functions
}
\references{
"Theory and experiment of InGaAsP and AlInGaAs long-wavelength strained quantum-well lasers" J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999
(\href{http://ieeexplore.ieee.org/document/760325/}{IEEE})
}
\keyword{datasets}
