% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1b-data.R
\docType{data}
\name{fig1b}
\alias{fig1b}
\title{InGaAsP Strain vs Bandgap Eg}
\format{\code{"data.frame"}}
\source{
J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999
}
\usage{
fig1b
}
\description{
Relation between InGaAsP bandgap and strain for composition x,y digitized from Fig. 1(b)
Data was digitized using Engauge Digitizer from Fig. 1(b) to compare against R functions
}
\examples{
data("fig1b")
head(fig1b)
}
\references{
"Theory and experiment of InGaAsP and AlInGaAs long-wavelength strained quantum-well lasers" J.Minch, S.H. Park, T. Keating and S.L. Chuang: IEEE JQE vol. 35 no. 5, 1999
(\href{http://ieeexplore.ieee.org/document/760325/}{IEEE})
}
\keyword{datasets}
